pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-email-triage"
version = "1.0.0"
description = "Enterprise email triage OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic>=2.10,<3",
  "PyYAML==6.0.2",
  "httpx==0.27.2",
  "python-dotenv>=1.0.1",
  "openai>=2.7.2",
  "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["env*"]
